requirements.txt
torch>=1.12.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-learn>=1.0.0
wfdb>=4.0.0
PyYAML>=6.0
tqdm>=4.62.0
tensorboard>=2.8.0